+++ src/components/anchor-navigation/anchor-section-divider.component.ts
@@ -1,10 +1,11 @@
 import styled from "styled-components";
 
-const StyledAnchorDivider = styled.div.attrs({
+const AnchorSectionDivider = styled.div.attrs({
   "data-element": "anchor-navigation-divider",
 })`
   background-color: var(--colorsActionMinor200);
   height: 1px;
 `;
 
-export default StyledAnchorDivider;
+AnchorSectionDivider.displayName = "AnchorSectionDivider";
+export default AnchorSectionDivider;